+++ extensions/iam/decentralized-identity/identity-did-core/src/main/java/org/eclipse/dataspaceconnector/iam/did/IdentityDidCoreExtension.java
@@ -49,14 +49,6 @@ public class IdentityDidCoreExtension implements ServiceExtension {
     @Inject
     private WebService webService;
 
-    public IdentityDidCoreExtension() {
-    }
-
-    public IdentityDidCoreExtension(IdentityHubStore hubStore, WebService webService) {
-        this.hubStore = hubStore;
-        this.webService = webService;
-    }
-
     @Override
     public String name() {
         return "Identity Did Core";